Update as of 07/31/2021

+++ delete-class.php
@@ -12,27 +12,38 @@
     redirect_to("view-teachers-and-rooms.php");
   }
 
+
+  $data_exist = return_duplicate_entry("schedule_block","class_id",$class_id,"",$connection);
+
+  if ($data_exist > 0) {
+    echo "<script type='text/javascript'>";
+    echo "alert('Error! Cannot delete class info. This class has schedule created. Delete the schedule first.');";
+    echo "</script>";
+    $URL="sections-and-classes.php";
+    echo "<script>location.href='$URL'</script>";
+  }
+  else{ 
     $query  = "DELETE FROM classes WHERE class_id ='".$class_id."' LIMIT 1";
     $result = mysqli_query($connection, $query);
 
     $query_schedule  = "DELETE FROM schedule_block WHERE class_id ='".$class_id."'";
     $result_schedule = mysqli_query($connection, $query_schedule);
 
-  //close database connection after an sql command
+    //close database connection after an sql command
 
-  if ($result === TRUE) {
-    echo "<script type='text/javascript'>";
-    echo "alert('Delete class successful!');";
-    echo "</script>";
+    if ($result === TRUE) {
+      echo "<script type='text/javascript'>";
+      echo "alert('Delete class successful!');";
+      echo "</script>";
 
-    $URL="classes.php";
-    echo "<script>location.href='$URL'</script>";
-  } else {
-    echo "Error updating record: " . $connection->error;
+      $URL="classes.php";
+      echo "<script>location.href='$URL'</script>";
+    } else {
+      echo "Error updating record: " . $connection->error;
+    }
+
+    if(isset($connection)){ mysqli_close($connection); }
   }
-
-  if(isset($connection)){ mysqli_close($connection); }
-
 ?>
