+++ delete-section.php
@@ -12,11 +12,18 @@
     redirect_to("view-teachers-and-rooms.php");
   }
 
+  $data_exist = return_duplicate_entry("sections","sec_id",$sec_id,"",$connection);
+
+  if ($data_exist > 0) {
+    echo "<script type='text/javascript'>";
+    echo "alert('Error! Cannot delete section info. This section is added to classes.');";
+    echo "</script>";
+    $URL="manage-sections.php";
+    echo "<script>location.href='$URL'</script>";
+  }
+  else{
     $query  = "DELETE FROM sections WHERE sec_id = {$sec_id} LIMIT 1";
     $result = mysqli_query($connection, $query);
-
-
-  //close database connection after an sql command
 
   if ($result === TRUE) {
     echo "<script type='text/javascript'>";
@@ -25,11 +32,14 @@
 
     $URL="manage-sections.php";
     echo "<script>location.href='$URL'</script>";
-  } else {
+    }
+    else {
     echo "Error updating record: " . $connection->error;
-  }
+    }
 
-  if(isset($connection)){ mysqli_close($connection); }
+    if(isset($connection)){ mysqli_close($connection); }
+
+  }
 
 ?>
